+++ sdk/documentintelligence/azure-ai-documentintelligence/pom.xml
@@ -38,6 +38,7 @@
       --add-opens com.azure.core/com.azure.core.implementation.util=ALL-UNNAMED
       --add-opens com.azure.ai.documentintelligence/com.azure.ai.documentintelligence=ALL-UNNAMED
     </javaModulesSurefireArgLine>
+    <jacoco.min.branchcoverage>0.29</jacoco.min.branchcoverage>
     <parallelizeLiveTests>false</parallelizeLiveTests>
     <graalvm.disable>false</graalvm.disable>
   </properties>
