+++ sdk/documentintelligence/azure-ai-documentintelligence/src/samples/java/com/azure/ai/documentintelligence/AnalyzeInvoices.java
@@ -32,7 +32,7 @@ public class AnalyzeInvoices {
      */
     public static void main(final String[] args) throws IOException {
         // Instantiate a client that will be used to call the service.
-        DocumentAnalysisClient client = new DocumentAnalysisClientBuilder()
+        DocumentIntelligenceClient client = new DocumentIntelligenceClientBuilder()
                                           .credential(new AzureKeyCredential("{key}"))
                                           .endpoint("https://{endpoint}.cognitiveservices.azure.com/")
                                           .buildClient();
