+++ sdk/documentintelligence/azure-ai-documentintelligence/src/samples/java/com/azure/ai/documentintelligence/administration/package-info.java
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 
 /**
- * Package containing sample classes for interacting with {@link com.azure.ai.documentintelligence.DocumentModelAdministrationAsyncClient} to
+ * Package containing sample classes for interacting with {@link com.azure.ai.documentintelligence.DocumentIntelligenceAdministrationAsyncClient} to
  * perform operations on Azure Form Recognizer
  */
 package com.azure.ai.documentintelligence.administration;
